:lipstick: applied layout to popup & added typescript support
