+++ src/containers/Document.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { ChakraProvider, Container } from '@chakra-ui/react';
+import { ChakraProvider } from '@chakra-ui/react';
 
 /**
  * @description
