+++ src/containers/layout/Popup.tsx
@@ -1,0 +1,8 @@
+import { Box } from '@chakra-ui/react';
+import React from 'react';
+
+const LayoutPopup: React.FC = (props) => {
+  return <Box>{props.children}</Box>;
+};
+
+export default LayoutPopup;
