+++ src/containers/layout/index.tsx
@@ -1,1 +1,3 @@
 export { default as NewtabLayout } from './Newtab';
+export { default as OptionsLayout } from './Options';
+export { default as PopupLayout } from './Popup';
