--- src/pages/Options/Options.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import './Options.css';
-
-interface Props {
-  title: string;
-}
-
-const Options: React.FC<Props> = ({ title }: Props) => {
-  return <div className="OptionsContainer">{title.toUpperCase()} PAGE</div>;
-};
-
-export default Options;
